provide list of files to ignore in docker
This is to ensure a more idempotent build context for docker. Without
this, you get installation errors due to different installation bindings
that arise when running npm install on the host OS. This is because the
node_modules folder gets copied to the container as part of the build
context, among other things.